requirements.txt
tensorflow
streamlit
fastapi
uvicorn
pillow
python-multipart
requests
